requirements.txt
Pillow>=9.0.0 
pandas>=1.3.0 
xlwings>=0.27.0 
